[refactor] 유형 이름 조회 함수 분리
- 기존 리뷰에 있었던 것처럼 History에서 해당 함수를 분리하여 util로 이동하였음
